Extract DiviK base class
