+++ divik/cluster/_divik/_dunn_sklearn.py
@@ -1,27 +1,12 @@
-from functools import partial
-from typing import Tuple
-
 import numpy as np
-import pandas as pd
-import scipy.spatial.distance as dist
-from sklearn.base import BaseEstimator, ClusterMixin, TransformerMixin
-from sklearn.utils.validation import check_is_fitted
-import tqdm
 
+from ._base import DiviKBase
 from ._dunn import dunn_divik
 from ._report import DivikReporter
-import divik.feature_selection as fs
 from divik.cluster import _kmeans as km
-import divik._summary as summary
-from divik._utils import (
-    normalize_rows,
-    DivikResult,
-    context_if,
-    maybe_pool,
-)
 
 
-class DunnDiviK(BaseEstimator, ClusterMixin, TransformerMixin):
+class DunnDiviK(DiviKBase):
     """DiviK clustering
 
     Parameters
@@ -194,108 +179,15 @@ class DunnDiviK(BaseEstimator, ClusterMixin, TransformerMixin):
                  n_jobs: int = None,
                  random_seed: int = 0,  # TODO: Rework to use RandomState
                  verbose: bool = False):
-        self.gap_trials = gap_trials
-        self.distance_percentile = distance_percentile
-        self.max_iter = max_iter
-        self.distance = distance
-        self.minimal_size = minimal_size
-        self.rejection_size = rejection_size
-        self.rejection_percentage = rejection_percentage
-        self.minimal_features_percentage = minimal_features_percentage
-        self.features_percentage = features_percentage
+        super().__init__(gap_trials, distance_percentile, max_iter, distance,
+                         minimal_size, rejection_size, rejection_percentage,
+                         minimal_features_percentage, features_percentage,
+                         k_max, normalize_rows, use_logfilters, filter_type,
+                         keep_outliers, n_jobs, random_seed, verbose)
         self.fast_kmeans_iter = fast_kmeans_iter
-        self.k_max = k_max
-        self.normalize_rows = normalize_rows
-        self.use_logfilters = use_logfilters
-        self.filter_type = filter_type
-        self.keep_outliers = keep_outliers
-        self.n_jobs = n_jobs
-        self.random_seed = random_seed
-        self.verbose = verbose
         self._validate_arguments()
-
-    def _validate_arguments(self):
-        self._validate_clustering()
-        self._validate_feature_selection()
-
-    def _validate_feature_selection(self):
-        if self.minimal_features_percentage < 0 \
-                or self.minimal_features_percentage > 1:
-            raise ValueError('minimal_features_percentage must be in range'
-                             ' [0, 1]')
-        if self.features_percentage < 0 or self.features_percentage > 1:
-            raise ValueError('features_percentage must be in range [0, 1]')
-        if self.features_percentage < self.minimal_features_percentage:
-            raise ValueError('features_percentage must be higher than or equal'
-                             ' to minimal_features_percentage')
-        if self.filter_type not in ['gmm', 'outlier', 'auto', 'none']:
-            raise ValueError(
-                "filter_type must be in ['gmm', 'outlier', 'auto', 'none']")
-
-    def _validate_clustering(self):
-        if self.gap_trials <= 0:
-            raise ValueError('gap_trials must be greater than 0')
-        if self.distance_percentile < 0 or self.distance_percentile > 100:
-            raise ValueError('distance_percentile must be in range [0, 100]')
-        if self.max_iter <= 0:
-            raise ValueError('max_iter must be greater than 0')
-        if self.minimal_size is not None and self.minimal_size < 0:
-            raise ValueError('minimal_size must be greater or equal to 0')
         if self.fast_kmeans_iter > self.max_iter or self.fast_kmeans_iter < 0:
             raise ValueError('fast_kmeans_iter must be in range [0, max_iter]')
-
-    def fit(self, X, y=None):
-        """Compute DiviK clustering.
-
-        Parameters
-        ----------
-        X : array-like or sparse matrix, shape=(n_samples, n_features)
-            Training instances to cluster. It must be noted that the data
-            will be converted to C ordering, which will cause a memory
-            copy if the given data is not C-contiguous.
-        y : Ignored
-            not used, present here for API consistency by convention.
-        """
-        if np.isnan(X).any():
-            raise ValueError("NaN values are not supported.")
-
-        with context_if(self.verbose, tqdm.tqdm, total=X.shape[0]) as progress:
-            self.result_ = self._divik(X, progress)
-
-        self.labels_, self.paths_ = summary.merged_partition(self.result_,
-                                                             return_paths=True)
-        self.reverse_paths_ = {
-            value: key for key, value in self.paths_.items()}
-        self.filters_ = np.array(
-            [self._get_filter(path) for path in self.reverse_paths_],
-            dtype=bool)
-        self.centroids_ = pd.DataFrame(X).groupby(self.labels_, sort=True)\
-            .mean().values
-        self.depth_ = summary.depth(self.result_)
-        self.n_clusters_ = summary.total_number_of_clusters(self.result_)
-
-        return self
-
-    def _get_rejection_size(self, X):
-        rejection_size = 0
-        if self.rejection_size is not None:
-            rejection_size = max(rejection_size, self.rejection_size)
-        if self.rejection_percentage is not None:
-            rejection_size = max(
-                rejection_size, int(self.rejection_percentage * X.shape[0]))
-        return rejection_size
-
-    def _get_filter(self, path):
-        """This method extracts features filter used for each centroid"""
-        result = self.result_
-        for item in path[:-1]:
-            result = result.subregions[item]
-        return result.feature_selector.selected_
-
-    def _needs_normalization(self):
-        if self.normalize_rows is None:
-            return self.distance == 'correlation'
-        return self.normalize_rows
 
     def _fast_kmeans(self):
         single_kmeans = km.KMeans(
@@ -320,27 +212,6 @@ class DunnDiviK(BaseEstimator, ClusterMixin, TransformerMixin):
             verbose=self.verbose)
         return kmeans
 
-    def _gmm_filter(self):
-        return fs.HighAbundanceAndVarianceSelector(
-            use_log=self.use_logfilters,
-            min_features_rate=self.minimal_features_percentage)
-
-    def _outlier_filter(self):
-        return fs.OutlierAbundanceAndVarianceSelector(
-            use_log=self.use_logfilters,
-            min_features_rate=self.minimal_features_percentage,
-            p=self.features_percentage)
-
-    def _feature_selector(self, n_features):
-        if (self.filter_type == 'auto' and n_features > 250) \
-                or self.filter_type == 'gmm':
-            return self._gmm_filter()
-        if self.filter_type == 'auto' or self.filter_type == 'outlier':
-            return self._outlier_filter()
-        if self.filter_type == 'none':
-            return fs.NoSelector()
-        raise ValueError("Unknown filter type: %s" % self.filter_type)
-
     def _divik(self, X, progress):
         fast = self._fast_kmeans()
         full = self._full_kmeans()
@@ -356,118 +227,3 @@ class DunnDiviK(BaseEstimator, ClusterMixin, TransformerMixin):
             feature_selector=self._feature_selector(X.shape[1]),
             minimal_size=minimal_size, rejection_size=rejection_size,
             report=report)
-
-    def fit_predict(self, X, y=None):
-        """Compute cluster centers and predict cluster index for each sample.
-
-        Convenience method; equivalent to calling fit(X) followed by
-        predict(X).
-
-        Parameters
-        ----------
-
-        X : {array-like, sparse matrix}, shape = [n_samples, n_features]
-            New data to transform.
-
-        y : Ignored
-            not used, present here for API consistency by convention.
-
-        Returns
-        -------
-
-        labels : array, shape [n_samples,]
-            Index of the cluster each sample belongs to.
-        """
-        return self.fit(X).labels_
-
-    def fit_transform(self, X, y=None, **fit_params):
-        """Compute clustering and transform X to cluster-distance space.
-
-        Equivalent to fit(X).transform(X), but more efficiently implemented.
-
-        Parameters
-        ----------
-
-        X : {array-like, sparse matrix}, shape = [n_samples, n_features]
-            New data to transform.
-
-        y : Ignored
-            not used, present here for API consistency by convention.
-
-        Returns
-        -------
-
-        X_new : array, shape [n_samples, self.n_clusters_]
-            X transformed in the new space.
-        """
-        return self.fit(X).transform(X)
-
-    def transform(self, X):
-        """Transform X to a cluster-distance space.
-
-        In the new space, each dimension is the distance to the cluster
-        centers.  Note that even if X is sparse, the array returned by
-        `transform` will typically be dense.
-
-        Parameters
-        ----------
-
-        X : {array-like, sparse matrix}, shape = [n_samples, n_features]
-            New data to transform.
-
-        Returns
-        -------
-
-        X_new : array, shape [n_samples, self.n_clusters_]
-            X transformed in the new space.
-        """
-        check_is_fitted(self)
-        if self._needs_normalization():
-            X = normalize_rows(X)
-        distances = np.hstack([
-            dist.cdist(
-                X[:, selector], centroid[np.newaxis, selector], self.distance)
-            for selector, centroid in zip(self.filters_, self.centroids_)
-        ])
-        return distances
-
-    def predict(self, X):
-        """Predict the closest cluster each sample in X belongs to.
-
-        In the vector quantization literature, `cluster_centers_` is called
-        the code book and each value returned by `predict` is the index of
-        the closest code in the code book.
-
-        Parameters
-        ----------
-
-        X : {array-like, sparse matrix}, shape = [n_samples, n_features]
-            New data to predict.
-
-        Returns
-        -------
-
-        labels : array, shape [n_samples,]
-            Index of the cluster each sample belongs to.
-        """
-        check_is_fitted(self)
-        if self._needs_normalization():
-            X = normalize_rows(X)
-        predict = partial(_predict_path, result=self.result_)
-        with maybe_pool(self.n_jobs) as pool:
-            paths = pool.map(predict, X)
-        labels = [self.reverse_paths_[path] for path in paths]
-        return np.array(labels, dtype=np.int32)
-
-
-def _predict_path(observation: np.ndarray, result: DivikResult) -> Tuple[int]:
-    path = []
-    observation = observation[np.newaxis, :]
-    division = result
-    while division is not None:
-        local_X = division.feature_selector.transform(observation)
-        label = int(division.clustering.predict(local_X))
-        path.append(label)
-        division = division.subregions[label]
-    path = tuple(path)
-    return path
